BB-5449: Change inventory threshold field type to decimal.

+++ src/Oro/Bundle/InventoryBundle/Form/Extension/CategoryInventoryThresholdFormExtension.php
@@ -7,7 +7,7 @@ use Symfony\Component\Form\FormBuilderInterface;
 
 use Oro\Bundle\CatalogBundle\Form\Type\CategoryType;
 use Oro\Bundle\EntityBundle\Form\Type\EntityFieldFallbackValueType;
-use Oro\Bundle\ValidationBundle\Validator\Constraints\Integer;
+use Oro\Bundle\ValidationBundle\Validator\Constraints\Decimal;
 
 class CategoryInventoryThresholdFormExtension extends AbstractTypeExtension
 {
@@ -30,7 +30,7 @@ class CategoryInventoryThresholdFormExtension extends AbstractTypeExtension
             [
                 'label' => 'oro.inventory.inventory_threshold.label',
                 'value_options' => [
-                    'constraints' => [new Integer()]
+                    'constraints' => [new Decimal()]
                 ]
             ]
         );
